refactor(cypher): split lowering projection + reentry runtime helpers
